Move validationCriteria and visualisationConfig schema check to importRows
