+++ packages/sync-server/app/admin/importSchemas/questionSchemas.js
@@ -1,6 +1,7 @@
 import * as yup from 'yup';
 import { SurveyScreenComponent, baseValidationShape, baseConfigShape } from './baseSchemas';
-import { configString, validationString } from './jsonString';
+import { configString, validationString, visualisationConfigString } from './jsonString';
+import { isNumberOrFloat } from '../../utils/numbers';
 
 const columnReferenceConfig = baseConfigShape.shape({
   column: yup.string().required(),
@@ -73,21 +74,53 @@ const numberConfig = baseConfigShape.shape({
   unit: yup.string(),
   rounding: yup.number(),
 });
+
+const normalRangeObjectSchema = yup
+  .object()
+  .shape({
+    min: yup.number(),
+    max: yup.number(),
+    ageUnit: yup.string().oneOf(['years', 'months', 'weeks']),
+    ageMin: yup.number(),
+    ageMax: yup.number(),
+  })
+  .noUnknown()
+  .test({
+    name: 'normalRange',
+    message: ctx => `normalRange should have either min or max, got ${JSON.stringify(ctx.value)}`,
+    test: value => {
+      if (!value) {
+        return true;
+      }
+      return isNumberOrFloat(value.min) || isNumberOrFloat(value.max);
+    },
+  });
+
+const visualisationConfigSchema = yup.object().shape({
+  yAxis: yup.object().shape({
+    graphRange: yup
+      .object()
+      .shape({
+        min: yup.number().required(),
+        max: yup.number().required(),
+      })
+      .required(),
+    interval: yup.number().required(),
+  }),
+});
+
 const numberValidationCriteria = baseValidationShape.shape({
   min: yup.number(),
   max: yup.number(),
-  normalRange: yup
-    .object()
-    .shape({
-      min: yup.number(),
-      max: yup.number(),
-    })
-    .noUnknown(),
+  normalRange: yup.lazy(value =>
+    Array.isArray(value) ? yup.array().of(normalRangeObjectSchema) : normalRangeObjectSchema,
+  ),
 });
 
 export const SSCNumber = SurveyScreenComponent.shape({
   config: configString(numberConfig),
   validationCriteria: validationString(numberValidationCriteria),
+  visualisationConfig: visualisationConfigString(visualisationConfigSchema),
 });
 export const SSCCalculatedQuestion = SurveyScreenComponent.shape({
   config: configString(numberConfig),
