+++ packages/sync-server/app/utils/numbers.js
@@ -1,0 +1,7 @@
+export const isNumberOrFloat = value => {
+  if (typeof value !== 'number') {
+    return false;
+  }
+
+  return !isNaN(value);
+};
